refactored not to use navigation directory
