+++ .github/CODEOWNERS
@@ -9,10 +9,8 @@
 
 /assets                                         @fuxingloh @thedoublejay
 /constants                                      @fuxingloh @thedoublejay
-
 /components                                     @fuxingloh @thedoublejay
 /hooks                                          @fuxingloh @thedoublejay
-/navigation                                     @fuxingloh @thedoublejay
 /screens                                        @fuxingloh @thedoublejay
 /store                                          @fuxingloh @thedoublejay
