+++ .github/labels.yml
@@ -60,8 +60,6 @@
 - color: fbca04
   name: area/assets
 - color: fbca04
-  name: area/navigation
-- color: fbca04
   name: area/playground
 - color: fbca04
   name: area/translations
