+++ screens/BottomTabNavigator/index.tsx
@@ -2,10 +2,10 @@ import { NavigationContainer, DefaultTheme, DarkTheme } from '@react-navigation/
 import { createStackNavigator } from '@react-navigation/stack'
 import * as React from 'react'
 
-import { NotFoundScreen } from '../screens/NotFoundScreen'
+import { NotFoundScreen } from '../NotFoundScreen'
 import { BottomTabNavigator } from './BottomTabNavigator'
 import { LinkingConfiguration } from './LinkingConfiguration'
-import { useColorScheme } from '../hooks/design/useColorScheme'
+import { useColorScheme } from '../../hooks/design/useColorScheme'
 
 // If you are not familiar with React Navigation, we recommend going through the
 // "Fundamentals" guide: https://reactnavigation.org/docs/getting-started
