+++ screens/BottomTabNavigator/screens/BalancesScreen/BalancesScreen.test.tsx
