+++ screens/BottomTabNavigator/screens/LiquidityScreen/LiquidityScreen.tsx
@@ -3,11 +3,11 @@ import tailwind from 'tailwind-rn'
 import { createStackNavigator } from '@react-navigation/stack'
 import { useDispatch, useSelector } from 'react-redux'
 
-import { Text, View } from '../../components/Themed'
-import { translate } from '../../translations'
+import { Text, View } from '../../../../components/Themed'
+import { translate } from '../../../../translations'
 import { Button } from 'react-native'
-import { decrement, increment, incrementAsync, incrementIfOdd } from '../../store/liquidity'
-import { RootState } from '../../store'
+import { decrement, increment, incrementAsync, incrementIfOdd } from '../../../../store/liquidity'
+import { RootState } from '../../../../store'
 
 export function LiquidityScreen (): JSX.Element {
   const count = useSelector<RootState>(state => state.counter.value)
