+++ screens/Main.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Navigation } from '../navigation'
+import { Navigation } from './BottomTabNavigator'
 import { StatusBar } from 'expo-status-bar'
 import { SafeAreaProvider } from 'react-native-safe-area-context'
