+++ screens/Main.web.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import tailwind from 'tailwind-rn'
 import { StyleSheet, View } from 'react-native'
-import { Navigation } from '../navigation'
+import { Navigation } from './BottomTabNavigator'
 import { PlaygroundNavigator } from '../playground/Playground'
 
 export function Main (): JSX.Element {
