+++ screens/WalletNavigator/WalletScreen.tsx
